fixes #46 : [bug][seeg] type 'serial' should be mapped to integer

+++ seeg-maven-plugin/src/main/java/com/sporniket/libre/javabeans/seeg/ExtractionWorkspacePostgresql.java
@@ -49,7 +49,7 @@ public class ExtractionWorkspacePostgresql extends ExtractionWorkspaceBase imple
 							"bigserial", "long"
 					},
 					{
-							"serial", "long"
+							"serial", "integer"
 					},
 					{
 							"int8", "long"
